requirement.txt
alembic
blinker
click
colorama
dnspython
email-validator
Flask
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
greenlet
Flask-Cors
Flask-Bcrypt
idna
importlib-metadata
itsdangerous
Jinja2
Mako
MarkupSafe
PyJWT
PyMySQL
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
zipp
gunicorn
